+++ spec/mongoid/criteria_spec.rb
@@ -4105,6 +4105,25 @@ describe Mongoid::Criteria do
     end
   end
 
+  describe "#with" do
+
+    let!(:criteria) do
+      Band.where(name: "Depeche Mode").with(collection: "artists")
+    end
+
+    after do
+      Band.persistence_options.clear
+    end
+
+    it "retains the criteria selection" do
+      criteria.selector.should eq("name" => "Depeche Mode")
+    end
+
+    it "sets the persistence options" do
+      Band.persistence_options.should eq(collection: "artists")
+    end
+  end
+
   describe "#within_box" do
 
     before do
@@ -4198,6 +4217,7 @@ describe Mongoid::Criteria do
     end
 
     it "returns the matching documents" do
+      p criteria.to_a
       criteria.should eq([ match ])
     end
   end